Working Version
Allows the user to place Pieces in the board by selecting the piece
from the hand, and choosing the coordinates on the board to place the
in. Scoring is not set up yet. Changed the output message for the
column cause it kept confusing me. Nice work by the group